Refactor delegate routes MM-357
