+++ conrtrollers/delegate/index.js
@@ -1,0 +1,11 @@
+const { read } = require('./read');
+const { create } = require('./create');
+const { remove } = require('./remove');
+const { update } = require('./update');
+
+module.exports = {
+  read,
+  create,
+  remove,
+  update,
+};
