+++ routes/DelegateRoutes.js
@@ -1,14 +1,9 @@
-const router = require("express").Router();
-const ResponseHandler = require("./utils/ResponseHandler");
+/* eslint-disable max-len */
+const router = require('express').Router();
 
-const DelegateService = require("../services/DelegateService");
-const BlockchainService = require("../services/BlockchainService");
-
-const Validator = require("./utils/Validator");
-const Constants = require("../constants/Constants");
-const Register = require("../models/Register");
-const { getCleanedDid } = require("./utils/DidClean");
-const Messages = require("../constants/Messages");
+const delegate = require('../conrtrollers/delegate');
+const Validator = require('./utils/Validator');
+const Constants = require('../constants/Constants');
 
 /**
  * @openapi
@@ -24,31 +19,23 @@ const Messages = require("../constants/Messages");
  *     responses:
  *       200:
  *         description: Puede devolver ok o error en algun parametro
- *       401: 
+ *       401:
  *         description: Acción no autorizada
  *       500:
  *         description: Error interno del servidor
  */
 
 router.get(
-	"/all",
-	Validator.validate([
-		{
-			name: "token",
-			validate: [Constants.USER_TYPES.Read_Delegates],
-			isHead: true
-		}
-	]),
-	Validator.checkValidationResult,
-	async function (_, res) {
-		try {
-			const delegates = await DelegateService.getAll();
-			const result = delegates.map(({ did, name }) => ({ did, name }));
-			return ResponseHandler.sendRes(res, result);
-		} catch (err) {
-			return ResponseHandler.sendErr(res, err);
-		}
-	}
+  '/all',
+  Validator.validate([
+    {
+      name: 'token',
+      validate: [Constants.USER_TYPES.Read_Delegates],
+      isHead: true,
+    },
+  ]),
+  Validator.checkValidationResult,
+  delegate.read,
 );
 
 /**
@@ -81,46 +68,32 @@ router.get(
  *     responses:
  *       200:
  *         description: Puede devolver ok o error en algun parametro
- *       401: 
+ *       401:
  *         description: Acción no autorizada
  *       500:
  *         description: Error interno del servidor
  */
 
 router.post(
-	"/",
-	Validator.validate([
-		{
-			name: "token",
-			validate: [Constants.USER_TYPES.Write_Delegates],
-			isHead: true
-		},
-		{ name: "name", validate: [Constants.VALIDATION_TYPES.IS_STRING] },
-		{
-			name: "did",
-			validate: [Constants.VALIDATION_TYPES.IS_STRING]
-		},
-		{
-			name: "registerId",
-			validate: [Constants.VALIDATION_TYPES.IS_STRING]
-		}
-	]),
-	Validator.checkValidationResult,
-	async function (req, res) {
-		try {
-			const { name, did, registerId } = req.body;
-
-			// autorizo en la blockchain
-			await BlockchainService.addDelegate(registerId, did);
-
-			// registro autorizacion en la bd local
-			await DelegateService.create(did, name, registerId);
-
-			return ResponseHandler.sendRes(res, { did, name });
-		} catch (err) {
-			return ResponseHandler.sendErr(res, err);
-		}
-	}
+  '/',
+  Validator.validate([
+    {
+      name: 'token',
+      validate: [Constants.USER_TYPES.Write_Delegates],
+      isHead: true,
+    },
+    { name: 'name', validate: [Constants.VALIDATION_TYPES.IS_STRING] },
+    {
+      name: 'did',
+      validate: [Constants.VALIDATION_TYPES.IS_STRING],
+    },
+    {
+      name: 'registerId',
+      validate: [Constants.VALIDATION_TYPES.IS_STRING],
+    },
+  ]),
+  Validator.checkValidationResult,
+  delegate.create,
 );
 
 /**
@@ -147,41 +120,27 @@ router.post(
  *     responses:
  *       200:
  *         description: Puede devolver ok o error en algun parametro
- *       401: 
+ *       401:
  *         description: Acción no autorizada
  *       500:
  *         description: Error interno del servidor
  */
 
 router.delete(
-	"/",
-	Validator.validate([
-		{
-			name: "token",
-			validate: [Constants.USER_TYPES.Write_Delegates],
-			isHead: true
-		},
-		{
-			name: "did",
-			validate: [Constants.VALIDATION_TYPES.IS_STRING]
-		}
-	]),
-	Validator.checkValidationResult,
-	async function (req, res) {
-		try {
-			const { did } = req.body;
-
-			// revoco autorizacion en la blockchain
-			await BlockchainService.removeDelegate(did);
-
-			// registro revocacion en la bd local
-			const { name } = await DelegateService.delete(did);
-
-			return ResponseHandler.sendRes(res, { did, name });
-		} catch (err) {
-			return ResponseHandler.sendErr(res, err);
-		}
-	}
+  '/',
+  Validator.validate([
+    {
+      name: 'token',
+      validate: [Constants.USER_TYPES.Write_Delegates],
+      isHead: true,
+    },
+    {
+      name: 'did',
+      validate: [Constants.VALIDATION_TYPES.IS_STRING],
+    },
+  ]),
+  Validator.checkValidationResult,
+  delegate.remove,
 );
 
 /**
@@ -205,30 +164,20 @@ router.delete(
  *     responses:
  *       200:
  *         description: Puede devolver ok o error en algun parametro
- *       401: 
+ *       401:
  *         description: Acción no autorizada
  *       500:
  *         description: Error interno del servidor
  */
 
 router.post(
-	"/didDelegationValid",
-	Validator.validate([
-		{ name: "didDelegate", validate: [Constants.VALIDATION_TYPES.IS_STRING] },
-		{ name: "registerId", validate: [Constants.VALIDATION_TYPES.IS_STRING] }
-	]),
-	Validator.checkValidationResult,
-	async function (req, res) {
-		try {
-			const { didDelegate, registerId } = req.body;
-
-			// seteo el nombre en la blockchain
-			const result = await BlockchainService.validDelegate(registerId, didDelegate);
-			return ResponseHandler.sendRes(res, result);
-		} catch (err) {
-			return ResponseHandler.sendErr(res, err);
-		}
-	}
+  '/didDelegationValid',
+  Validator.validate([
+    { name: 'didDelegate', validate: [Constants.VALIDATION_TYPES.IS_STRING] },
+    { name: 'registerId', validate: [Constants.VALIDATION_TYPES.IS_STRING] },
+  ]),
+  Validator.checkValidationResult,
+  delegate.update,
 );
 
 module.exports = router;
